Replaced "create.js" special-handling of autocreate by standard events
Historically, exhibit relied on use of a special "create.js" file
added at the end of the scripts to trigger actual (automatic) creation
of the exhibit once all scripts were loaded.  I've incorporated this
automatic creation into the new event-based workflow for exhibit:

1. Moved autocreation code to exhibit.js, named Exhibit.createFromDOM.
This is an imperfect name but configureFromDOM was already taken;
createFromDOM intializes and loads the db while configureFromDOM only
configures the visualization atop an existing db
2. Defined a new event, staticConfigured.exhibit, that is triggered at
the end of static configuration in final.js .  Static configuration
covers loading of scripts but also configuration of static components
(e.g. registering importers) that are relied on when building an exhibit
3. If autoCreate is specified, final.js sets a listener to triger on
staticConfiguration.  Setting this listener would more appropriately
go in exhibit-api.js where the autoCreate parameter is defined
relevant, but jQuery hasn't yet been loaded there so I can't use it to
set the listener
4. Removed create.js, which is no longer relevant